Break up PrinterOptions
