+++ src/index.ts
@@ -6,7 +6,9 @@ export * from './Printers/Communication/UsbPrinterDeviceChannel';
 export * from './Printers/Configuration/PrinterOptions';
 export * from './Printers/Languages/EplPrinterCommandSet';
 export * from './Printers/Languages/PrinterCommandSet';
+export * from './Printers/Models/EplPrinterModels';
 export * from './Printers/Models/PrinterModel';
+export * from './Printers/Models/PrinterModelDb';
 export * from './Printers/Printer';
 export * from './PrinterUsbManager';
 export * from './WebZlpError';
